+++ src/behave/spot.h
@@ -57,7 +57,7 @@ public:
     void setLocation(SpotFireLocation::SpotFireLocationEnum location);
     void setRidgeToValleyDistance(double ridgeToValleyDistance, LengthUnits::LengthUnitsEnum ridgeToValleyDistanceUnits);
     void setRidgeToValleyElevation(double ridgeToValleyElevation, LengthUnits::LengthUnitsEnum elevationUnits);
-    void setTorchingTrees(double torchingTrees);
+    void setTorchingTrees(int torchingTrees);
     void setTreeHeight(double treeHeight, LengthUnits::LengthUnitsEnum  treeHeightUnits);
     void setTreeSpecies(SpotTreeSpecies::SpotTreeSpeciesEnum treeSpecies);
     void setWindSpeedAtTwentyFeet(double windSpeedAtTwentyFeet, SpeedUnits::SpeedUnitsEnum windSpeedUnits);
@@ -72,7 +72,7 @@ public:
         SpeedUnits::SpeedUnitsEnum windSpeedUnits, double flameLength, LengthUnits::LengthUnitsEnum flameLengthUnits);
 	void updateSpotInputsForTorchingTrees(SpotFireLocation::SpotFireLocationEnum location, double ridgeToValleyDistance,
         LengthUnits::LengthUnitsEnum ridgeToValleyDistanceUnits, double ridgeToValleyElevation, LengthUnits::LengthUnitsEnum elevationUnits,
-        double downwindCoverHeight, LengthUnits::LengthUnitsEnum coverHeightUnits, double torchingTrees, double DBH,
+        double downwindCoverHeight, LengthUnits::LengthUnitsEnum coverHeightUnits, int torchingTrees, double DBH,
         LengthUnits::LengthUnitsEnum DBHUnits, double treeHeight, LengthUnits::LengthUnitsEnum  treeHeightUnits,
         SpotTreeSpecies::SpotTreeSpeciesEnum treeSpecies, double windSpeedAtTwentyFeet, SpeedUnits::SpeedUnitsEnum windSpeedUnits);
 
@@ -84,7 +84,7 @@ public:
     SpotFireLocation::SpotFireLocationEnum getLocation();
     double getRidgeToValleyDistance(LengthUnits::LengthUnitsEnum ridgeToValleyDistanceUnits);
     double getRidgeToValleyElevation(LengthUnits::LengthUnitsEnum elevationUnits);
-    double getTorchingTrees();
+    int getTorchingTrees();
     double getTreeHeight(LengthUnits::LengthUnitsEnum  treeHeightUnits);
     SpotTreeSpecies::SpotTreeSpeciesEnum getTreeSpecies();
     double getWindSpeedAtTwentyFeet(SpeedUnits::SpeedUnitsEnum windSpeedUnits);
